clean up Milestone V2 API
[#188514709]
