+++ tracker/api/serializers.py
@@ -800,6 +800,7 @@ class MilestoneSerializer(
     WithPermissionsSerializerMixin, EventNestedSerializerMixin, TrackerModelSerializer
 ):
     type = ClassNameField()
+    event = EventSerializer()
 
     class Meta:
         model = Milestone
